Move AssertionUnit from the extension to messaging and adjust all cmake files
