+++ Source/com/CMakeLists.txt
@@ -51,7 +51,6 @@ target_link_libraries(${TARGET}
         PRIVATE
           ${NAMESPACE}Core::${NAMESPACE}Core
           ${NAMESPACE}Messaging::${NAMESPACE}Messaging
-          ${NAMESPACE}Assertion::${NAMESPACE}Assertion
           CompileSettingsDebug::CompileSettingsDebug
         )
