--- Source/extensions/assertion/AssertionUnit.h
@@ -1,46 +1,0 @@
-/*
- * If not stated otherwise in this file or this component's LICENSE file the
- * following copyright and licenses apply:
- *
- * Copyright 2020 Metrological
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include "Module.h"
-
-namespace Thunder {
-namespace Assertion {
-
-    class EXTERNAL AssertionUnit : public IAssertionUnit {
-    private:
-        AssertionUnit(const AssertionUnit&) = delete;
-        AssertionUnit& operator=(const AssertionUnit&) = delete;
-        AssertionUnit(AssertionUnit&&) = delete;
-        AssertionUnit& operator=(AssertionUnit&&) = delete;
-
-    protected:
-        AssertionUnit();
-
-    public:
-        ~AssertionUnit() override;
-
-    public:
-        static AssertionUnit& Instance();
-
-        void AssertionEvent(Core::Messaging::IStore::Assert& metadata, const Core::Messaging::TextMessage& message) override;
-    };
-}
-}
